Added async JSRoot loding (#38)
* Added async JSRoot loding

This time loading doesn't come with vulnerabilities

* Moved project to parent folder

* removing warnings
